refactor(dwarf,elfo,personagem): vida em double nao int, ajustados parametros refactor(ElfoNoturno): ajustado como elfo noturno perde vida

+++ src/modulo-01-java-OO/projeto-lotr-bluej/ElfoTest.java
@@ -31,7 +31,7 @@ public class ElfoTest
         Dwarf d1 = new Dwarf("d1");
         int numFlInic = e1.getFlechas();
         int numExpInic = e1.getExp();
-        int vidaDwarfInic = d1.getVida();
+        double vidaDwarfInic = d1.getVida();
         e1.atirarFlechaDwarfs(d1);
         boolean result = (numFlInic > e1.getFlechas() 
                             && numExpInic < e1.getExp()
